refactor: rearrange file structure
