+++ Server/docker-compose.yml
@@ -12,6 +12,5 @@ services:
       - "5432:5432"
     volumes:
       - db_data:/var/lib/postgresql/data
-
 volumes:
   db_data:
