+++ client/src/components/home/heroSection.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 export default function HeroSection() {
   return (
     <section className="flex flex-col items-center justify-center h-screen w-full text-center px-6 relative overflow-hidden">
-      <div className="max-w-4xl mx-auto p-8 bg-white/10 backdrop-blur-md rounded-2xl shadow-lg">
+      <div className="max-w-4xl mx-auto p-8 bg-white/10 backdrop-blur-md rounded-2xl shadow-lg mb-10">
         <h1 className="text-4xl font-extrabold mb-4 leading-tight">
           Embark on an Unforgettable Journey of Discovery and Adventure!
         </h1>
